requirements.txt
yaspin==2.2.0
requests==2.32.5
groq==1.0.0
python-dotenv==1.2.1